Move out some types from RPCClientProxyHandler to node-api in preparation for moving the AMQP scheme
